refactoring names for ActionButton
